火警立即执行,放弃执行的逻辑调整优化

+++ OpticalValleyUnite/Classes/UI/FireControlProfession/Model/YQFireLocationModel.swift
@@ -23,7 +23,7 @@ class YQFireLocationModel: NSObject {
      },
      */
     
-    var firePointId : Int = -1
+    var firePointId : Int64 = -1
     
     var name : String = ""
